refactor: inline layout engine types, rename to processRoomActions and use reduce
